Remove deprecated device_pool_drivers api; created V1 API base controller
