+++ app/controllers/api/v1/trips_controller.rb
@@ -1,4 +1,4 @@
-class API::V1::TripsController < API::ApiController
+class API::V1::TripsController < API::V1::BaseController
   before_action :authenticate_customer
   before_action :get_trip, :validate_trip_customer_match, except: [:create]
